--- DataTrans/DataTrans.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30225.117
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DataTrans", "DataTrans.vcxproj", "{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Debug|x64.ActiveCfg = Debug|x64
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Debug|x64.Build.0 = Debug|x64
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Debug|x86.ActiveCfg = Debug|Win32
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Debug|x86.Build.0 = Debug|Win32
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Release|x64.ActiveCfg = Release|x64
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Release|x64.Build.0 = Release|x64
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Release|x86.ActiveCfg = Release|Win32
-		{D1EAB85D-0A00-45C5-9CBD-F8E26783C153}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BB9A2C0A-AFC1-4489-ABF7-598354E89108}
-	EndGlobalSection
-EndGlobal
