--- DataTrans/readme.md
@@ -1,19 +1,0 @@
-# DataTrans branch
-
-## 描述
-
-虽然叫DataTrans, 但其实这是整个MVVM框架
-
-由于view和GUI事件关系比较密切，暂时不打算在这个工程里实现view
-
-view model会用到qt中的某些库，而且用到这些库的地方一般就是GUI逻辑，因此主要实现非逻辑部分，完善view model和model的沟通
-
-希望不要破坏我的格式
-
-## 测试方法
-
-按下F5
-
-## 当前进度
-
-进行中
