--- DataTrans/viewmodel.cpp
@@ -1,1 +1,0 @@
-#include "viewmodel.h"
